Issue #000 fix: de-dup enhancement to handle it properly.
